use/display BTC/KB & DCR/KB values in fee rate limit field

+++ client/asset/dcr/config.go
@@ -38,7 +38,7 @@ type Config struct {
 	RPCCert          string  `ini:"rpccert"`
 	UseSplitTx       bool    `ini:"txsplit"`
 	FallbackFeeRate  float64 `ini:"fallbackfee"`
-	FeeRateLimit     uint64  `ini:"feeratelimit"`
+	FeeRateLimit     float64 `ini:"feeratelimit"`
 	RedeemConfTarget uint64  `ini:"redeemconftarget"`
 	// Context should be canceled when the application exits. This will cause
 	// some cleanup to be performed during shutdown.
